Add tx2gene mapping for un-annotated transcripts
Add a manual tx2gene mapping un-annotated transcripts, taking the
transcript IDs as gene; a fully connected tx2gene mapping for the
transcriptome to be used by Alevin is required, hence this addition.

Also, merge the separate processes for building the tx2gene mapping and
adding the mitochondrion genome to it and the transcriptome FASTA, for
process efficiency.